Added tests for basic rendering on each major acmcsuf.com page
