+++ src/routes/blog/index.test.ts
@@ -1,0 +1,15 @@
+import { beforeEach, test, expect } from 'vitest';
+import { cleanup, render } from '@testing-library/svelte';
+
+import Blog from './index.svelte';
+
+beforeEach(cleanup);
+
+test('can render', () => {
+  render(Blog);
+});
+
+test('can find the correct page title', () => {
+  const { getByText } = render(Blog);
+  expect(getByText('The official acmCSUF blog.')).toBeDefined();
+});
