+++ src/routes/paths/index.svelte
